MIBE-63: fix tests + readme
