+++ frontend/src/side_panel/mixed_initiative_interaction/metacognitive_support/levelOfSupportUtils.ts
@@ -2,19 +2,19 @@ import { LevelOfSupport } from '../../models/support_and_MII/UserSupport';
 
 const LoSDescription: Record<LevelOfSupport, string> = {
   text: 'The system will only provide instructions, but will indicate which steps have been completed.',
-  overlay:
+  hints:
     'The system will provide instructions and show you what you need to interact with by scrolling to it and putting a pink rectangle around it.',
-  click: 'The system will perform the next action for you automatically.',
+  auto: 'The system will perform the next action for you automatically.',
 };
 
 const increaseLevelOfSupport = (level: LevelOfSupport): LevelOfSupport => {
   switch (level) {
     case 'text':
-      return 'overlay';
-    case 'overlay':
-      return 'click';
-    case 'click':
-      return 'click';
+      return 'hints';
+    case 'hints':
+      return 'auto';
+    case 'auto':
+      return 'auto';
   }
 };
 
@@ -22,10 +22,10 @@ const decreaseLevelOfSupport = (level: LevelOfSupport): LevelOfSupport => {
   switch (level) {
     case 'text':
       return 'text';
-    case 'overlay':
+    case 'hints':
       return 'text';
-    case 'click':
-      return 'overlay';
+    case 'auto':
+      return 'hints';
   }
 };
